Adding more feature v1.1

+++ app/Http/Controllers/PostController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Category;
 use App\Models\Post;
 use Illuminate\Http\Request;
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\File;
 
 class PostController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -23,6 +24,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Post::class);
         $categories = Category::all();
         return view('create',compact('categories'));
     }
@@ -32,6 +34,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Post::class);
+
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', 'max:255'],
@@ -67,6 +71,8 @@ class PostController extends Controller
     public function edit(string $id)
     {
         $post = Post::find($id);
+        $this->authorize('update', $post);
+
         $categories = Category::all();
         return view('edit',compact('post','categories'));
     }
@@ -76,13 +82,15 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $post = Post::findOrFail($id);
+        $this->authorize('update', $post);
+
         $request->validate([
             'title' => ['required', 'max:255'],
             'category_id' => ['required', 'integer'],
             'description' => ['required']
         ],['category_id.required' => "The category field is required."]);
 
-        $post = Post::findOrFail($id);
 
         if($request->hasFile('image')){
             $request->validate([
@@ -108,18 +116,23 @@ class PostController extends Controller
     public function destroy(string $id)
     {
         $post = Post::findOrFail($id);
+        $this->authorize('delete', $post);
+
         $post->delete();
 
         return redirect()->back();
     }
 
     public function trashed(){
+        $this->authorize('trash', Post::class);
+
         $posts = Post::onlyTrashed()->paginate(10);
         return view('trashed',compact('posts'));
     }
 
     public function restore($id){
         $post = Post::onlyTrashed()->findOrFail($id);
+        $this->authorize('restore', $post);
         $post->restore();
 
         return redirect()->back();
@@ -127,6 +140,8 @@ class PostController extends Controller
 
     public function forceDelete($id){
         $post = Post::onlyTrashed()->findOrFail($id);
+        $this->authorize('forceDelete', $post);
+
         File::delete(public_path($post->image));
         $post->forceDelete();
